Refactoring some of the common region handling
