+++ spinnaker_graph_front_end/examples/Conways/partitioned_example_a_no_vis_no_buffer/conways_basic_cell.py
@@ -13,11 +13,12 @@ from spinn_front_end_common.utilities.constants import SYSTEM_BYTES_REQUIREMENT
 from spinn_front_end_common.utilities.exceptions import ConfigurationException
 from spinn_front_end_common.utilities.helpful_functions \
     import locate_memory_region_for_placement
-from spinn_front_end_common.interface.simulation import simulation_utilities
 from spinn_front_end_common.abstract_models.impl \
     import MachineDataSpecableVertex
 
 from spinnaker_graph_front_end.utilities import SimulationBinary
+from spinnaker_graph_front_end.utilities.data_utils \
+    import generate_system_data_region
 
 # general imports
 from enum import Enum
@@ -60,14 +61,11 @@ class ConwayBasicCell(
             self, spec, placement, machine_graph, routing_info, iptags,
             reverse_iptags, machine_time_step, time_scale_factor,
             n_machine_time_steps):
-
-        # Setup words + 1 for flags + 1 for recording size
-        setup_size = SYSTEM_BYTES_REQUIREMENT
+        # Generate the system data region for simulation .c requirements
+        generate_system_data_region(spec, self.DATA_REGIONS.SYSTEM.value,
+                                    self, machine_time_step, time_scale_factor)
 
         # reserve memory regions
-        spec.reserve_memory_region(
-            region=self.DATA_REGIONS.SYSTEM.value,
-            size=setup_size, label='systemInfo')
         spec.reserve_memory_region(
             region=self.DATA_REGIONS.TRANSMISSIONS.value,
             size=self.TRANSMISSION_DATA_SIZE, label="inputs")
@@ -79,13 +77,7 @@ class ConwayBasicCell(
             size=8, label="neighour_states")
         spec.reserve_memory_region(
             region=self.DATA_REGIONS.RESULTS.value,
-            size=(n_machine_time_steps * 4) + 4, label="results")
-
-        # simulation .c requirements
-        spec.switch_write_focus(self.DATA_REGIONS.SYSTEM.value)
-        spec.write_array(simulation_utilities.get_simulation_header_array(
-            self.get_binary_file_name(), machine_time_step,
-            time_scale_factor))
+            size=(n_machine_time_steps + 1) * 4, label="results")
 
         # check got right number of keys and edges going into me
         partitions = \
