+++ spinnaker_graph_front_end/examples/template/template_vertex.py
@@ -7,8 +7,6 @@ from spinn_front_end_common.utilities import globals_variables
 from spinn_front_end_common.utilities.constants import SYSTEM_BYTES_REQUIREMENT
 from spinn_front_end_common.utilities.helpful_functions \
     import locate_memory_region_for_placement, read_config_int
-from spinn_front_end_common.interface.simulation.simulation_utilities \
-    import get_simulation_header_array
 from spinn_front_end_common.abstract_models.impl \
     import MachineDataSpecableVertex
 from spinn_front_end_common.interface.buffer_management.buffer_models\
@@ -17,6 +15,8 @@ from spinn_front_end_common.interface.buffer_management\
     import recording_utilities
 
 from spinnaker_graph_front_end.utilities import SimulationBinary
+from spinnaker_graph_front_end.utilities.data_utils \
+    import generate_system_data_region
 
 from enum import Enum
 import logging
@@ -41,11 +41,10 @@ class TemplateVertex(
                ('RECORDED_DATA', 2)])
 
     def __init__(self, label, constraints=None):
-
-        self._recording_size = 5000
-
         MachineVertex.__init__(self, label=label, constraints=constraints)
         SimulationBinary.__init__(self, "c_template_vertex.aplx")
+
+        self._recording_size = 5000
 
         config = globals_variables.get_simulator().config
         self._buffer_size_before_receive = None
@@ -87,44 +86,21 @@ class TemplateVertex(
         :param reverse_iptags: the collection of reverse iptags generated by\
                 the tag allocator
         """
+        # Note the placement now that we know it!
         self.placement = placement
 
-        # Create the data regions
-        self._reserve_memory_regions(spec)
+        # Generate the system data region for simulation .c requirements
+        generate_system_data_region(spec, self.DATA_REGIONS.SYSTEM.value,
+                                    self, machine_time_step, time_scale_factor)
 
-        # write simulation interface data
-        spec.switch_write_focus(self.DATA_REGIONS.SYSTEM.value)
-        spec.write_array(get_simulation_header_array(
-            self.get_binary_file_name(), machine_time_step,
-            time_scale_factor))
-
-        # write recording data interface
-        spec.switch_write_focus(self.DATA_REGIONS.RECORDED_DATA.value)
-        spec.write_array(recording_utilities.get_recording_header_array(
-            [self._recording_size], self._time_between_requests,
-            self._buffer_size_before_receive, iptags))
-
-        # Get the key, assuming all outgoing edges use the same key
-        has_key = 0
-        key = routing_info.get_first_key_from_pre_vertex(self, PARTITION_ID)
-        if key is None:
-            key = 0
-        else:
-            has_key = 1
-
-        # Write the transmission region
-        spec.switch_write_focus(self.DATA_REGIONS.TRANSMISSION.value)
-        spec.write_value(has_key)
-        spec.write_value(key)
+        # Generate the application data regions
+        self._reserve_app_memory_regions(spec)
+        self._write_app_memory_regions(spec, routing_info, iptags)
 
         # End-of-Spec:
         spec.end_specification()
 
-    def _reserve_memory_regions(self, spec):
-        spec.reserve_memory_region(
-            region=self.DATA_REGIONS.SYSTEM.value,
-            size=SYSTEM_BYTES_REQUIREMENT,
-            label='systemInfo')
+    def _reserve_app_memory_regions(self, spec):
         spec.reserve_memory_region(
             region=self.DATA_REGIONS.TRANSMISSION.value,
             size=self.TRANSMISSION_REGION_N_BYTES, label="transmission")
@@ -132,6 +108,21 @@ class TemplateVertex(
             region=self.DATA_REGIONS.RECORDED_DATA.value,
             size=recording_utilities.get_recording_header_size(1),
             label="recording")
+
+    def _write_app_memory_regions(self, spec, routing_info, iptags):
+        # Get the key, assuming all outgoing edges use the same key
+        key = routing_info.get_first_key_from_pre_vertex(self, PARTITION_ID)
+
+        # Write the transmission region
+        spec.switch_write_focus(self.DATA_REGIONS.TRANSMISSION.value)
+        spec.write_value(int(key is not None))
+        spec.write_value(0 if key is None else key)
+
+        # write recording data interface
+        spec.switch_write_focus(self.DATA_REGIONS.RECORDED_DATA.value)
+        spec.write_array(recording_utilities.get_recording_header_array(
+            [self._recording_size], self._time_between_requests,
+            self._buffer_size_before_receive, iptags))
 
     def read(self, placement, buffer_manager):
         """ Get the recorded data
